Separate schemas into invidual files, and move them to central location in parent business-form directory. rename some items
